+++ tests/test_debug/snapshot_main_frame.py
@@ -7,7 +7,7 @@ This module is intended exclusively for testing purposes and should not be used 
 
 import json
 
-from tests.test_debug.snapshot import snapshot
+from tests.test_debug.test_snapshot import snapshot
 
 # globally defined variables
 team_lead = "David Liu"
